% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.JMcmprsk.R
\name{anova.JMcmprsk2}
\alias{anova.JMcmprsk2}
\title{ANOVA of longitudinal model}
\usage{
\method{anova}{JMcmprsk2}(object, coeff = c("beta", "gamma", "alpha"),
  ...)
}
\arguments{
\item{object}{The JMcmprsk2 object returned by either jmo or jmc function.}

\item{coeff}{Types of coefficients selected for anova. Note "alpha" is only avaiable to jmo type JMcmprsk2 object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Return a Wald test statistic and the p value
  \tabular{ll}{
      \code{beta}    \tab  The Wald test for fixed effects for the longitutal part,i.e.  \eqn{\beta} in jmo or jmc output. \cr
      \code{gamma}    \tab  The Wald test for fixed effects for the survival part,i.e.  \eqn{\gamma} in jmo or jmc output. "gamma1" stands for test for competing risk 1 and "gamma2" stands for test for competing risk 2  \cr
      \code{alpha}    \tab  The Wald test for non-proportional odds covariates,i.e.  \eqn{\alpha} in jmo output. \cr
  }
}
\description{
Joint modelling for longitutal and censored data with competing risks
}
