% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.JMcmprsk.R
\name{print.JMcmprsk2}
\alias{print.JMcmprsk2}
\title{Print JMcmprsk2}
\usage{
\method{print}{JMcmprsk2}(x, ...)
}
\arguments{
\item{x}{Object of class 'JMcmprsk2'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of JMcmprsk2 object.
}
\seealso{
\code{\link{jmc}}
}
\author{
Hong Wang
}
